clean-up/bootstrap scripts finished

+++ Project/S3Handler.py
@@ -25,27 +25,6 @@
  #############################################################################
 import boto3
 
-
-POLICY_TEMPLATE = """{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Effect": "Allow",
-      "Action": ["s3:ListBucket"],
-      "Resource": ["arn:aws:s3:::{BUCKET_NAME}"]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "s3:PutObject",
-        "s3:GetObject",
-        "s3:DeleteObject"
-      ],
-      "Resource": ["arn:aws:s3:::{BUCKET_NAME}/*"]
-    }
-  ]
-}"""
-
 def get_random_prefix(is_public=False):
     """ Generates a new random bucket name with a prefix 'snovak.'
 
@@ -81,11 +60,7 @@ class S3Handler(object):
         # acquire a new bucket name if no bucket name was specified
         if self.bucket_name is None:
             self.bucket_name = get_random_prefix(is_public=True)
-        policy = POLICY_TEMPLATE.replace("{BUCKET_NAME}", self.bucket_name)    
-        bucket_policy = self.s3_resource.BucketPolicy(self.bucket_name)
-        bucket_policy.policy = policy
-        self.bucket =  bucket_policy.Bucket()
-        
+        self.bucket = self.s3_resource.create_bucket(Bucket=self.bucket_name, ACL='public-read')
 
     def create_new_bucket(self):
         # acquire a new bucket name if no bucket name was specified
@@ -93,18 +68,24 @@ class S3Handler(object):
             self.bucket_name = get_random_prefix()
         self.bucket = self.s3_resource.create_bucket(Bucket=self.bucket_name)
 
-    def upload_file(self, src, s3_key):
+    def upload_file(self, src, s3_key, public=False):
         """Attempts to upload specified file to current bucket
 
         Args:
             src (str): path to source file to be uploaded
             s3_key (str): value to be used as key in destination bucket
+            public (bool): if true, file will be made available to public, false by default
         Returns:
             None
         Raises:
             S3HandlerError if field 'bucket' is not initialized
         """
         self.bucket.upload_file(src, s3_key)
+        if public:
+            # apply ACL , grant read public for everyone
+            self.bucket.upload_file(src, s3_key, ExtraArgs = {'ACL': 'public-read'})
+        else:
+            self.bucket.upload_file(src, s3_key)
 
     def download_file(self, s3_key, dst):
         """Attempts to download specified file using s3_key and save it as 'dst'
@@ -130,8 +111,23 @@ class S3Handler(object):
             S3HandlerError if field 'bucket' is not initialized
         """
         self.bucket.delete_objects(Delete={'Objects': [{'Key': s3_key}]})
-    
+
+    def is_bucket_created(self):
+        from dateutil.parser import parse
+        try:
+            parse(self.bucket.creation_date)
+            return True
+        except:
+            return False
+
     def delete_bucket(self):
+        # empty whole bucket
+        if not self.is_bucket_created():
+            print("Empty bucket, nothing to delete")
+            return
+        for bucket_object in self.bucket.objects.all():
+            print("Deleting object:" + str(bucket_object.key))
+            bucket_object.delete()
         self.bucket.delete()
 
     def get_bucket_name(self):
@@ -140,6 +136,14 @@ class S3Handler(object):
 
 
 if __name__ == '__main__':
-    handler = S3Handler("bestnssnovakname.rtrk.bbt.2016.23.4123123213xxxxx2x")
-    handler.create_new_public_bucket()
-    handler.upload_file("log.txt", "log.txt")
+    # test public bucket / public file
+    """
+    EXISTING_BUCKET_NAME = "snovak.public.input.bucket.2016.1234567890xxxxxyyyzzz2"
+    s3_input_handle =  S3Handler(EXISTING_BUCKET_NAME)
+    s3_input_handle.create_new_public_bucket()
+    s3_input_handle.upload_file('test-file.txt', 'my-awesome-key', True)
+    """
+
+    s3_test_handle = S3Handler('snovak.public.input.bucket.2016.1234567890')
+    s3_test_handle.download_file('snovak.client.1482355937.7394884__test-file.txt', 'slasdasdasdasd.json')
+
